requirements.txt
django 
django-filter==2.0.0
django-heroku==0.3.1